BasicEntityPersister::executeInserts() -- use stable ordering of columns.
BasicEntityPersister::getInsertSQL() used unnumbered parameters and
relies on BEP::getInsertColumnList() to use the same ordering as
BEP::prepareInsertData(). This need not be the case, as
BEP::prepareInsertData() just calls BEP::prepareUpdateData().

Just make sure that the same ordering of columns is used.

In order to really reuse getInsertColumnList() we need an option to
generate unquoted insert columns, so we have now

getInsertColumnList(bool $unquoted = false)

which is used in executeInserts() with a truth value.